Workaround so the WorkUnitURLStreamHandlerProvider is loaded by the service loader.
It seems that the URL way of doing it is not working when the classloader loading the URL class and the classloader loading the WorkUnitURLStreamHandlerProvider is different.
I'm still not sure why this is (not) working like that, so the workaound will keep us going with very small changes.
The ChainURLStreamHandlerProvider is initialised in a static way in the WorkUnitService class as this is where we need it. We could consider initialising it somewhere else if needed.